Specify the start and end of a schedule

Creates a schedule of events occuring before and/or after a certain date or other scheduled event.

after(start_event, within_given = NULL)

before(end_event, within_given = NULL)

in_between(start_event, end_event, within_given = NULL)

on_or_after(start_event, within_given = NULL)

on_or_before(end_event, within_given = NULL)

Arguments

start_event, end_event
The start and/or end events of the schedule. Can be either:

A date-time object or

A schedule.

within_given
Required only when either start_date or end_date is a schedule. Can be either:

A string shortcut for a period type: either "day", "week", "month", "quarter" or "year". Or,

A date accessor function. Eg. lubridate::day().

Value

A schedule of events occuring before and/or after the events specified.

Details

When start_date and/or end_date are date-time objects they place fixed lower and upper limits on the resulting schedule; for more details see the vignette on understanding schedule limits.

Examples

christmas <-  only_occurring(in_month("Dec"), on_mday(25))
new_years_eve <-  only_occurring(in_month("Dec"), on_mday(31))

after_christmas <- after(christmas, within_given = lubridate::year)

schedule_days(after_christmas, from = 2000, to = 2001)

#>  [1] "2000-12-26" "2000-12-27" "2000-12-28" "2000-12-29" "2000-12-30"
#>  [6] "2000-12-31" "2001-12-26" "2001-12-27" "2001-12-28" "2001-12-29"
#> [11] "2001-12-30" "2001-12-31"

in_between_christmas_and_new_years_eve <-
    in_between(christmas,
               new_years_eve,
               within_given = lubridate::year)

schedule_days(in_between_christmas_and_new_years_eve, from = 2000, to = 2001)

#>  [1] "2000-12-26" "2000-12-27" "2000-12-28" "2000-12-29" "2000-12-30"
#>  [6] "2001-12-26" "2001-12-27" "2001-12-28" "2001-12-29" "2001-12-30"